refactor(sync): extract sync_engine/error.rs
Move SyncError, RecoveryStrategy, REWIND_DISTANCE, MAX_REWINDS_PER_RUN,
the classification helpers, the Display/Error impls, and all error-taxonomy
tests out of sync_engine/mod.rs into a dedicated error.rs submodule.

Re-exported from mod.rs via `pub(crate) use error::{...}`, so in-crate
users of `crate::wallet::sync_engine::SyncError` (and the other re-exported
names) continue to compile unchanged.

One test stays in mod.rs — sqlite_lock_contention_is_recognised — because
it still exercises the `is_sqlite_lock_contention` helper that lives in
mod.rs. A follow-up commit will move that helper (and this test) into the
lwd submodule.

23 tests pass; 11 of the 12 previously-inlined error-taxonomy tests now
run under `wallet::sync_engine::error::tests::*`.
